move Config struct to separate file
