+++ src/controller.rs
@@ -7,7 +7,7 @@ use crate::inputfile::{InputFile, InputFileReader};
 use crate::line_range::{LineRanges, RangeCheckResult};
 use crate::output::OutputType;
 use crate::printer::{InteractivePrinter, Printer, SimplePrinter};
-use crate::{Config, PagingMode};
+use crate::config::{Config, PagingMode};
 
 pub struct Controller<'a> {
     config: &'a Config<'a>,
